+++ tests/data/unformatted/401_explain_select.sql
@@ -1,0 +1,13 @@
+explain select top 25
+foooooooo, barrrrrrrrr, bazzzzzzzzzzzzzzzz, quxxxxxxxxxxx, foooooooooooooooooo, barrrrrrrrr
+from something
+)))))__SQLFMT_OUTPUT__(((((
+explain
+select top 25
+    foooooooo,
+    barrrrrrrrr,
+    bazzzzzzzzzzzzzzzz,
+    quxxxxxxxxxxx,
+    foooooooooooooooooo,
+    barrrrrrrrr
+from something